Cloud_Functions/Banco_Central_CF/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
pandas>=2.0.0
requests>=2.25.0
pyarrow>=14.0.0
beautifulsoup4>=4.10.0
lxml>=4.9.0
